Always run in /app and expect paths in /github/workspace
The CI suite here was passing implicitly because our `Gemfile` was
checked out when running the action "locally". Under real usage GitHub
changes the working directory to `/github/workspace`, so that won't
work.

I'm not sure the best way to do this. What I've done feels brittle, but
it works. Basically we use `CMD` to change to `/app` before doing
anything, and we hard-code a prefix of `/github/workspace` onto
`INPUT_PATH` so that we can find the original file to validate.